+++ index.d.ts
@@ -271,10 +271,7 @@ declare module "moleculer-web" {
 
 	// CorsOptions
 	// From: https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/cors/index.d.ts
-	type CustomOrigin = (
-		requestOrigin: string | undefined,
-		callback: (err: Error | null, allow?: boolean) => void
-	) => void;
+	type CustomOrigin = (origin: string) => boolean;
 
 	export interface CorsOptions {
 		origin?: boolean | string | RegExp | (string | RegExp)[] | CustomOrigin;